refactor(AvailableOffline): Move code to AvailableOfflineManager and DriveFileManager+AvailableOffline
